refactor(storybook): add interfaces file in storybook folder
