+++ packages/calcite-components/src/components/interfaces.ts
@@ -1,24 +1,15 @@
 /* Note: using `.d.ts` file extension will exclude it from the output build */
 export type Alignment = "start" | "center" | "end";
-export type Appearance = "solid" | "outline" | "outline-fill" | "transparent";
+export type Appearance = "outline" | "outline-fill" | "solid" | "transparent";
 export type Columns = 1 | 2 | 3 | 4 | 5 | 6;
 export type FlipContext = "both" | "start" | "end";
 export type Kind = "brand" | "danger" | "info" | "inverse" | "neutral" | "warning" | "success";
-export type Layout =
-  | "horizontal"
-  | "vertical"
-  | "grid"
-  | "inline"
-  | "center"
-  | "auto"
-  | "fixed"
-  | "none"
-  | "horizontal-single";
+export type Layout = "horizontal" | "vertical" | "grid";
 export type LogicalFlowPosition = "inline-start" | "inline-end" | "block-start" | "block-end";
 export type ModeClass = "calcite-mode-light" | "calcite-mode-dark" | "calcite-mode-auto";
 export type ModeName = "light" | "dark" | "auto";
-export type Position = "start" | "end" | "top" | "bottom";
-export type SelectionAppearance = "icon" | "border";
+export type Position = "start" | "end";
+export type SelectionAppearance = "border" | "icon";
 export type SelectionMode =
   | "single"
   | "none"
@@ -30,27 +21,3 @@ export type SelectionMode =
 export type Scale = "s" | "m" | "l";
 export type Status = "invalid" | "valid" | "idle";
 export type Width = "auto" | "half" | "full";
-export type ArrowType = "inline" | "edge" | "none";
-export type DisplayMode = "dock" | "float" | "overlay";
-export type ToggleDisplay = "button" | "switch";
-export type Dir = "ltr" | "rtl";
-export type ButtonType = "radio" | "checkbox";
-export type InteractionMode = "interactive" | "static";
-export type IconType = "chevron" | "caret" | "ellipsis" | "overflow";
-export type DeterminateType = "determinate" | "indeterminate";
-export type FillType = "single" | "range";
-export type LabelType = "percent" | "units";
-export type ClickType = "click" | "hover";
-export type CollapseDirection = "down" | "up";
-export type TextType =
-  | "text"
-  | "textarea"
-  | "email"
-  | "password"
-  | "tel"
-  | "number"
-  | "search"
-  | "file"
-  | "time"
-  | "date";
-export type Mode = "offset" | "name";
